refactor: the project structure (#166)
