+++ mlora/__init__.py
@@ -1,9 +1,11 @@
-from mlora.utils import convert_hf_to_pth, save_lora_model, setup_seed, setup_logging, setup_cuda_check
-from mlora.tokenizer import Tokenizer
-from mlora.model import LLMModel
-from mlora.model_llama import LlamaModel
-from mlora.model_chatglm import ChatGLMModel
-from mlora.modelargs import LLMModelArgs, MultiLoraBatchData, LoraBatchDataConfig
+from mlora.utils import (convert_hf_to_pth, save_lora_model,
+                         setup_seed, setup_logging, setup_cuda_check,
+                         load_base_model, init_lora_model)
+from mlora.tokenizer.tokenizer import Tokenizer
+from mlora.model.model import LLMModel
+from mlora.model.model_llama import LlamaModel
+from mlora.model.model_chatglm import ChatGLMModel
+from mlora.model.modelargs import LLMModelArgs, MultiLoraBatchData, LoraBatchDataConfig
 from mlora.dispatcher import TrainTask, Dispatcher
 
 __all__ = [
@@ -21,5 +23,7 @@ __all__ = [
     "save_lora_model",
     "setup_seed",
     "setup_logging",
-    "setup_cuda_check"
+    "setup_cuda_check",
+    "load_base_model",
+    "init_lora_model"
 ]
