+++ mlora/checkpoint/recompute.py
@@ -1,71 +1,10 @@
+from mlora.checkpoint.checkpoint import (check_backward_validity,
+                                         _get_autocast_kwargs,
+                                         get_device_states,
+                                         set_device_states,
+                                         detach_variable)
+
 import torch
-from typing import Any, Iterable, List, Tuple
-
-
-def detach_variable(inputs: Tuple[Any, ...]) -> Tuple[torch.Tensor, ...]:
-    if isinstance(inputs, tuple):
-        out = []
-        for inp in inputs:
-            if not isinstance(inp, torch.Tensor):
-                out.append(inp)
-                continue
-
-            x = inp.detach()
-            x.requires_grad = inp.requires_grad
-            out.append(x)
-        return tuple(out)
-    else:
-        raise RuntimeError("Only tuple tensors is supported.")
-
-
-def check_backward_validity(inputs: Iterable[Any]) -> None:
-    if not any(inp.requires_grad for inp in inputs if isinstance(inp, torch.Tensor)):
-        raise RuntimeError("Input need grad")
-
-
-def get_device_states(*args) -> Tuple[List[int], List[torch.Tensor]]:
-    fwd_gpu_devices = list({arg.get_device() for arg in args
-                            if isinstance(arg, torch.Tensor) and arg.is_cuda})
-
-    fwd_gpu_states = []
-    for device in fwd_gpu_devices:
-        with torch.cuda.device(device):
-            fwd_gpu_states.append(torch.cuda.get_rng_state())
-
-    return fwd_gpu_devices, fwd_gpu_states
-
-
-def set_device_states(devices, states) -> None:
-    for device, state in zip(devices, states):
-        with torch.cuda.device(device):
-            torch.cuda.set_rng_state(state)
-
-
-def _get_autocast_kwargs():
-    gpu_autocast_kwargs = {"enabled": torch.is_autocast_enabled(),
-                           "dtype": torch.get_autocast_gpu_dtype(),
-                           "cache_enabled": torch.is_autocast_cache_enabled()}
-
-    cpu_autocast_kwargs = {"enabled": torch.is_autocast_cpu_enabled(),
-                           "dtype": torch.get_autocast_cpu_dtype(),
-                           "cache_enabled": torch.is_autocast_cache_enabled()}
-
-    return gpu_autocast_kwargs, cpu_autocast_kwargs
-
-
-def pack_hook(to_offload: torch.Tensor) -> Tuple[torch.device, torch.Tensor]:
-    return to_offload.device, to_offload.to("cpu")
-
-
-def unpack_hook(to_offload_info: Tuple[torch.device, torch.Tensor]) -> torch.Tensor:
-    device, to_offload = to_offload_info
-    return to_offload.to(device)
-
-
-def CheckpointOffloadFunction(run_function, *args):
-    with torch.autograd.graph.saved_tensors_hooks(pack_hook, unpack_hook):
-        outputs = run_function(*args)
-    return outputs
 
 
 class CheckpointRecomputeFunction(torch.autograd.Function):
