+++ mlora/model/Embedding.py
@@ -1,0 +1,14 @@
+import torch
+import torch.nn.functional as F
+
+
+class Embedding(torch.nn.Module):
+    def __init__(self, embedding: torch.Tensor, pad_token: int):
+        super().__init__()
+        self.token_embedding_: torch.Tensor = embedding
+        self.padding_idx_: int = pad_token
+
+    def forward(self, tokens: torch.Tensor) -> torch.Tensor:
+        data = F.embedding(tokens, self.token_embedding_,
+                           padding_idx=self.padding_idx_)
+        return data
