+++ mlora/model/model.py
@@ -1,4 +1,4 @@
-from mlora.modelargs import MultiLoraBatchData, Masks
+from mlora.model.modelargs import MultiLoraBatchData, Masks
 
 import torch
 
@@ -84,23 +84,6 @@ def repeat_kv(x: torch.Tensor, n_rep: int) -> torch.Tensor:
         batch, n_kv_heads, n_rep, seq_len, head_dim)
     x = x.reshape(batch, n_kv_heads * n_rep, seq_len, head_dim)
     return x
-
-
-class RMSNorm(torch.nn.Module):
-    def __init__(self, weight: torch.Tensor, eps: float = 1e-6):
-        super().__init__()
-        self.norm_eps_ = eps
-        self.weight_ = weight
-
-    def _norm(self, data: torch.Tensor) -> torch.Tensor:
-        return data * torch.rsqrt(+ self.norm_eps_)
-
-    def forward(self, data: torch.Tensor) -> torch.Tensor:
-        input_dtype = data.dtype
-        v = data.to(torch.float32).pow(2).mean(-1, keepdim=True)
-        data = data * torch.rsqrt(v + self.norm_eps_)
-
-        return (self.weight_ * data).to(input_dtype)
 
 
 class LLMModel(metaclass=ABCMeta):
